backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.1
python-dotenv==1.0.1
bcrypt==4.1.2
psycopg2-binary==2.9.9